feat(migration): register IPC for migrating collections from BRU to YAML format
