+++ packages/bruno-electron/src/ipc/collection.js
@@ -73,12 +73,10 @@ const snapshotManager = require('../services/snapshot');
 const interpolateVars = require('./network/interpolate-vars');
 const { interpolateString } = require('./network/interpolate-string');
 const { getEnvVars, getTreePathFromCollectionToItem, mergeVars, parseBruFileMeta, hydrateRequestWithUuid, transformRequestToSaveToFilesystem } = require('../utils/collection');
-const { buildStampedFolderMigrateItems } = require('./migrate-folder-seq');
 const { getProcessEnvVars } = require('../store/process-env');
 const { getOAuth2TokenUsingAuthorizationCode, getOAuth2TokenUsingClientCredentials, getOAuth2TokenUsingPasswordCredentials, getOAuth2TokenUsingImplicitGrant, refreshOauth2Token } = require('../utils/oauth2');
 const { getCertsAndProxyConfig } = require('./network/cert-utils');
 const collectionWatcher = require('../app/collection-watcher');
-const { remount: remountCollectionV2 } = require('./mount');
 const { transformBrunoConfigBeforeSave, transformBrunoConfigAfterRead } = require('../utils/transformBrunoConfig');
 const { REQUEST_TYPES } = require('../utils/constants');
 const { cancelOAuth2AuthorizationRequest, isOauth2AuthorizationRequestInProgress } = require('../utils/oauth2-protocol-handler');
@@ -2679,176 +2677,6 @@ const registerRendererEventHandlers = (mainWindow, watcher) => {
         throw error;
       }
     } catch (error) {
-      throw error;
-    }
-  });
-
-  ipcMain.handle('renderer:migrate-collection-to-yml', async (event, collectionPathname, collectionUid) => {
-    const format = getCollectionFormat(collectionPathname);
-    if (format === 'yml') {
-      throw new Error('Collection is already in YML format');
-    }
-
-    // Stop the watcher during migration to avoid triggering events
-    if (watcher) {
-      watcher.removeWatcher(collectionPathname, mainWindow, collectionUid);
-    }
-
-    // Track all written yml files so we can roll back on failure
-    const writtenYmlFiles = [];
-
-    const tabPathMap = {};
-
-    try {
-      const brunoJsonPath = path.join(collectionPathname, 'bruno.json');
-      const brunoJsonContent = fs.readFileSync(brunoJsonPath, 'utf8');
-      const brunoConfig = JSON.parse(brunoJsonContent);
-
-      const collectionBruPath = path.join(collectionPathname, 'collection.bru');
-      let collectionRoot = {};
-      if (fs.existsSync(collectionBruPath)) {
-        const collectionBruContent = fs.readFileSync(collectionBruPath, 'utf8');
-        collectionRoot = parseCollection(collectionBruContent, { format: 'bru' });
-      }
-
-      const ymlBrunoConfig = { ...brunoConfig };
-      delete ymlBrunoConfig.version; // drop the bru format marker
-      ymlBrunoConfig.opencollection = '1.0.0';
-      // Carry the user-facing version: bru's collectionVersion becomes yml's info.version.
-      if (ymlBrunoConfig.collectionVersion) {
-        ymlBrunoConfig.version = ymlBrunoConfig.collectionVersion;
-      }
-      delete ymlBrunoConfig.collectionVersion;
-
-      const ocYmlPath = path.join(collectionPathname, 'opencollection.yml');
-      const ymlCollectionContent = stringifyCollection(collectionRoot, ymlBrunoConfig, { format: 'yml' });
-      await writeFile(ocYmlPath, ymlCollectionContent);
-      writtenYmlFiles.push(ocYmlPath);
-
-      const bruFiles = searchForFiles(collectionPathname, '.bru');
-      const envDirPath = path.join(collectionPathname, 'environments');
-      const bruFilesToDelete = [];
-
-      // Freeze sidebar folder order as contiguous seq on every visible folder before writing yml
-      const folderMigrateItems = buildStampedFolderMigrateItems(collectionPathname, bruFiles, (folderPath) => {
-        const folderBruPath = path.join(folderPath, 'folder.bru');
-        if (!fs.existsSync(folderBruPath)) {
-          return null;
-        }
-        const folderData = parseFolder(fs.readFileSync(folderBruPath, 'utf8'), { format: 'bru' });
-        return {
-          name: folderData?.meta?.name,
-          seq: folderData?.meta?.seq,
-          folderData
-        };
-      });
-
-      for (const item of folderMigrateItems) {
-        const ymlContent = stringifyFolder(item.folderData, { format: 'yml' });
-        const ymlFilePath = path.join(item.folderPath, 'folder.yml');
-        await writeFile(ymlFilePath, ymlContent);
-        writtenYmlFiles.push(ymlFilePath);
-        if (item.bruFilePath) {
-          bruFilesToDelete.push(item.bruFilePath);
-        }
-      }
-
-      for (const bruFilePath of bruFiles) {
-        const basename = path.basename(bruFilePath);
-        const dirname = path.dirname(bruFilePath);
-
-        if (basename === 'collection.bru' && path.normalize(dirname) === path.normalize(collectionPathname)) {
-          bruFilesToDelete.push(bruFilePath);
-          continue;
-        }
-
-        if (path.normalize(dirname) === path.normalize(envDirPath)) {
-          continue;
-        }
-
-        if (basename === 'folder.bru') {
-          continue;
-        }
-
-        const bruContent = fs.readFileSync(bruFilePath, 'utf8');
-        const requestData = parseRequest(bruContent, { format: 'bru' });
-        const ymlContent = stringifyRequest(requestData, { format: 'yml' });
-        const ymlFilePath = bruFilePath.replace(/\.bru$/, '.yml');
-        await writeFile(ymlFilePath, ymlContent);
-        moveRequestUid(bruFilePath, ymlFilePath);
-        tabPathMap[bruFilePath] = ymlFilePath;
-        writtenYmlFiles.push(ymlFilePath);
-        bruFilesToDelete.push(bruFilePath);
-      }
-
-      if (fs.existsSync(envDirPath)) {
-        const envBruFiles = searchForFiles(envDirPath, '.bru');
-        for (const envBruFilePath of envBruFiles) {
-          const envBruContent = fs.readFileSync(envBruFilePath, 'utf8');
-          const envData = parseEnvironment(envBruContent, { format: 'bru' });
-          const ymlContent = stringifyEnvironment(envData, { format: 'yml' });
-          const ymlFilePath = envBruFilePath.replace(/\.bru$/, '.yml');
-          await writeFile(ymlFilePath, ymlContent);
-          moveRequestUid(envBruFilePath, ymlFilePath);
-          writtenYmlFiles.push(ymlFilePath);
-          bruFilesToDelete.push(envBruFilePath);
-        }
-      }
-
-      for (const bruFile of bruFilesToDelete) {
-        fs.unlinkSync(bruFile);
-      }
-      fs.unlinkSync(brunoJsonPath);
-
-      try {
-        snapshotManager.remapCollectionTabPaths(collectionPathname, tabPathMap);
-      } catch (_) {
-      }
-
-      const { size, filesCount } = await getCollectionStats(collectionPathname);
-      ymlBrunoConfig.size = size;
-      ymlBrunoConfig.filesCount = filesCount;
-
-      try {
-        const remounted = await remountCollectionV2({ collectionUid, brunoConfig: ymlBrunoConfig });
-        if (!remounted && watcher) {
-          watcher.addWatcher(mainWindow, collectionPathname, collectionUid, ymlBrunoConfig, false, undefined, { ignoreInitial: true });
-        }
-      } catch (watcherError) {
-        console.error('Failed to re-attach watcher after migration:', watcherError);
-        try {
-          if (watcher) {
-            watcher.addWatcher(mainWindow, collectionPathname, collectionUid, ymlBrunoConfig, false, undefined, { ignoreInitial: true });
-          }
-        } catch (fallbackError) {
-          console.error('Fallback watcher attach failed after migration:', fallbackError);
-          mainWindow.webContents.send('main:display-error', {
-            message: `Collection migrated to yml, but live sync could not be re-enabled: ${fallbackError.message}. Please reopen the collection.`
-          });
-        }
-      }
-
-      return ymlBrunoConfig;
-    } catch (error) {
-      for (const ymlFile of writtenYmlFiles) {
-        try {
-          if (fs.existsSync(ymlFile)) {
-            fs.unlinkSync(ymlFile);
-          }
-        } catch (_) {
-        }
-      }
-
-      // Restart the watcher on the original bru collection
-      try {
-        const config = JSON.parse(fs.readFileSync(path.join(collectionPathname, 'bruno.json'), 'utf8'));
-        const remounted = await remountCollectionV2({ collectionUid, brunoConfig: config });
-        if (!remounted && watcher) {
-          watcher.addWatcher(mainWindow, collectionPathname, collectionUid, config);
-        }
-      } catch (watcherError) {
-        console.error('Failed to restart watcher after migration error:', watcherError);
-      }
       throw error;
     }
   });
